pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentr"
version = "0.1.0"
description = "AI Research Assistant with multi-agent architecture and conversation memory"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tavily-python>=0.3.0",
    "langfuse>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]
cli = [
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.scripts]
agentr = "cli_example:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentr*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
